fix: avoid warnings with MSVC
In both cases a `size_t` was being converted to a `uint32_t`. These
headers are used from application code, or at least from generated code,
and the application may be compiling with more warnings enabled than
normal.

+++ src/google/protobuf/string_block.h
@@ -145,7 +145,8 @@ inline StringBlock* StringBlock::Emplace(void* p, size_t n, StringBlock* next) {
   ABSL_DCHECK_EQ(n, NextSize(next));
   uint32_t doubled = static_cast<uint32_t>(n) * 2;
   uint32_t next_size = next ? std::min(doubled, max_size()) : min_size();
-  return new (p) StringBlock(next, false, RoundedSize(n), next_size);
+  return new (p) StringBlock(
+      next, false, RoundedSize(static_cast<uint32_t>(n)), next_size);
 }
 
 inline StringBlock* StringBlock::New(StringBlock* next) {
